refactor(SupportedAlgorithms): split class definition to separate file
