+++ src/SubtleCrypto.js
@@ -4,7 +4,7 @@
 const CryptoKey = require('./CryptoKey')
 const CryptoKeyPair = require('./CryptoKeyPair')
 const JsonWebKey = require('./JsonWebKey')
-const supportedAlgorithms = require('./algorithms/supportedAlgorithms')
+const supportedAlgorithms = require('./algorithms')
 const recognizedKeyUsages = require('./recognizedKeyUsages')
 const InvalidAccessError = require('./errors/InvalidAccessError')
 const NotSupportedError = require('./errors/NotSupportedError')
