+++ src/dictionaries/RsaHashedKeyAlgorithm.js
@@ -15,7 +15,7 @@ const CryptoKeyPair = require('../CryptoKeyPair')
 const JsonWebKey = require('../JsonWebKey')
 const KeyAlgorithm = require('./KeyAlgorithm')
 const RsaKeyAlgorithm = require('./RsaKeyAlgorithm')
-const supportedAlgorithms = require('../algorithms/supportedAlgorithms')
+const supportedAlgorithms = require('../algorithms')
 const DataError = require('../errors/DataError')
 const OperationError = require('../errors/OperationError')
 const InvalidAccessError = require('../errors/InvalidAccessError')
