+++ test/algorithms/supportedAlgorithmsSpec.js
@@ -11,8 +11,8 @@ chai.should()
 /**
  * Code under test
  */
-const supportedAlgorithms = require('../../src/algorithms/supportedAlgorithms')
-const SupportedAlgorithms = supportedAlgorithms.constructor
+const supportedAlgorithms = require('../../src/algorithms')
+const SupportedAlgorithms = require('../../src/algorithms/SupportedAlgorithms')
 const RegisteredAlgorithms = require('../../src/algorithms/RegisteredAlgorithms')
 const RsaHashedKeyAlgorithm = require('../../src/dictionaries/RsaHashedKeyAlgorithm')
 const ShaKeyAlgorithm = require('../../src/dictionaries/ShaKeyAlgorithm')
